requirements.txt
Django==3.2.8
djangorestframework==3.12.4
mysqlclient==2.0.3
python-dotenv==0.19.1
pytz==2021.3
djangorestframework-simplejwt==5.0.0
pyjwt==2.3.0
django-cors-headers==3.10.0
cloudinary==1.28.0
django-cors-headers==3.10.0
drf-yasg==1.20.0
python-dateutil==2.8.2